MAN-475 refactor route logic into controllers
