+++ server/@types/index.ts
@@ -5,3 +5,4 @@ export * from './Data.type'
 export * from './Appointment.type'
 export * from './ActivityLog.type'
 export * from './Option.type'
+export * from './Controller.type'
