--- server/controllers/renders/index.ts
@@ -1,7 +1,0 @@
-import { staffContacts } from './staffContacts'
-import { userScheduleController } from './caseload'
-
-export default {
-  staffContacts,
-  userScheduleController,
-}
