+++ server/middleware/index.ts
@@ -5,3 +5,4 @@ export * from './getSentences'
 export * from './getAppointment'
 export * from './redirectWizard'
 export * from './filterActivityLog'
+export * from './getPersonActivity'
