+++ server/routes/compliance.ts
@@ -1,50 +1,11 @@
 import { type Router } from 'express'
-import { auditService } from '@ministryofjustice/hmpps-audit-client'
-import { v4 } from 'uuid'
 import asyncMiddleware from '../middleware/asyncMiddleware'
 import type { Services } from '../services'
-import MasApiClient from '../data/masApiClient'
-import TierApiClient from '../data/tierApiClient'
 import type { Route } from '../@types'
-import ArnsApiClient from '../data/arnsApiClient'
-import { toPredictors, toRoshWidget } from '../utils/utils'
+import controllers from '../controllers'
 
 export default function complianceRoutes(router: Router, { hmppsAuthClient }: Services) {
   const get = (path: string | string[], handler: Route<void>) => router.get(path, asyncMiddleware(handler))
 
-  get('/case/:crn/compliance', async (req, res, _next) => {
-    const { crn } = req.params
-    const token = await hmppsAuthClient.getSystemClientToken(res.locals.user.username)
-
-    await auditService.sendAuditMessage({
-      action: 'VIEW_MAS_COMPLIANCE',
-      who: res.locals.user.username,
-      subjectId: crn,
-      subjectType: 'CRN',
-      correlationId: v4(),
-      service: 'hmpps-manage-people-on-probation-ui',
-    })
-
-    const arnsClient = new ArnsApiClient(token)
-    const masClient = new MasApiClient(token)
-    const tierClient = new TierApiClient(token)
-
-    const [personCompliance, tierCalculation, risks, predictors] = await Promise.all([
-      masClient.getPersonCompliance(crn),
-      tierClient.getCalculationDetails(crn),
-      arnsClient.getRisks(crn),
-      arnsClient.getPredictorsAll(crn),
-    ])
-
-    const risksWidget = toRoshWidget(risks)
-
-    const predictorScores = toPredictors(predictors)
-    res.render('pages/compliance', {
-      personCompliance,
-      tierCalculation,
-      crn,
-      risksWidget,
-      predictorScores,
-    })
-  })
+  get('/case/:crn/compliance', controllers.compliance.getCompliance(hmppsAuthClient))
 }
